% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{bp_pal}
\alias{bp_pal}
\alias{bp_ramp}
\title{Generate colour maps}
\usage{
bp_pal(
  type = "linear",
  option = 1L,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1
)

bp_ramp(
  n,
  type = "linear",
  option = "blue",
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1
)
}
\arguments{
\item{type}{A character string indicating the type of color map to use.
Three types are available:
\itemize{
\item \code{"linear"}
\item \code{"bipolar"}
\item \code{"multi"}
}}

\item{option}{A character string indicating the specific palette
Options are specific to type:
\itemize{
\item {
\code{"linear"}
\itemize{
\item \code{"blue"} (default)
\item \code{"green"}
\item \code{"coral"}
}
}
\item {
\code{"bipolar"}
\itemize{
\item \code{"green_coral"} (default)
\item \code{"blue_green"}
\item \code{"blue_coral"}
}
}
\item {
\code{"multi"}
\itemize{
\item \code{"fsc"} (default)
}
}
}}

\item{alpha}{The alpha transparency, a number in \link{0,1}, see argument alpha in
\code{\link[grDevices]{hsv}}.}

\item{begin}{The (corrected) hue in \link{0,1} at which the color map begins.}

\item{end}{The (corrected) hue in \link{0,1} at which the color map ends.}

\item{direction}{Sets the order of colors in the scale. If 1, the default,
colors are ordered from darkest to lightest. If -1, the order of colors is
reversed.}
}
\value{
\code{bp_pal} returns a function, which
accepts an integer \code{n} and produces a color ramp
with \code{n} values
}
\description{
Generate colour maps

Interpolate a colour ramp using a set of brand anchor points
}
