% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_blueprint}
\alias{scale_color_blueprint}
\alias{scale_colour_blueprint}
\alias{scale_fill_blueprint}
\title{Colour and fill scales using Blueprint-branded palettes}
\usage{
scale_color_blueprint(
  ...,
  type = "linear",
  option = "blue",
  discrete = FALSE,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1
)

scale_colour_blueprint(
  ...,
  type = "linear",
  option = "blue",
  discrete = FALSE,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1
)

scale_fill_blueprint(
  ...,
  type = "linear",
  option = "blue",
  discrete = FALSE,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1
)
}
\arguments{
\item{type}{A string}

\item{option}{A string}

\item{alpha}{A number \link{0, 1}.}

\item{begin}{A number \link{0, 1}.}

\item{end}{A number \link{0, 1}.}

\item{direction}{An integer: one of 1 (fwd), -1 (reverse)}
}
\description{
Colour and fill scales using Blueprint-branded palettes
}
