% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_blueprint_palettes.R
\name{show_blueprint_palettes}
\alias{show_blueprint_palettes}
\title{Display defined Palettes}
\usage{
show_blueprint_palettes()
}
\value{
a plot, showing all of the blueprint colour palettes.
}
\description{
Display defined Palettes
}
